refactor: application/application → application/usecase 폴더 평탄화
- 도메인명 'application'과 DDD Application Layer 명칭 충돌로 경로가 application/application/*로 이중 중첩되던 문제 해소
- 도메인 내부에서는 usecase/ 로 유스케이스 폴더 의미를 명시
- git mv로 히스토리 보존. 4개 파일의 import 경로 동반 업데이트
