+++ src/application/usecase/application-query.service.spec.ts
