+++ src/application/usecase/application-query.service.ts
